backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
requests
google-generativeai